abstract SDK_VERSION
