+++ packages/database/package.json
@@ -7,6 +7,7 @@
   "browser": "dist/index.cjs.js",
   "module": "dist/index.esm.js",
   "esm2017": "dist/index.esm2017.js",
+  "main-admin": "dist/index.admin.node.cjs.js",
   "files": [
     "dist"
   ],
