+++ packages/database/src/core/version.ts
@@ -1,0 +1,7 @@
+
+/** The semver (www.semver.org) version of the SDK. */
+export let SDK_VERSION = '';
+
+export function setSDKVersion(version: string): void {
+    SDK_VERSION = version;
+}
